Create Buy and Sell Stocks- K Transactions Allowed.java
Time Complexity :
Now that we have avoided the use of the innermost loop, we have the time complexity as O(kxn) where k is the number of transactions allowed and n is the number of days.

SPACE COMPLEXITY :
We have not done any reduction in the space, so the space complexity remains the same i.e. O(kxn).